+++ backend/src/production/production.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { ProductionController } from "./production.controller";
 import { ProductionService } from "./production.service";
-import { BlogDto, ProductionDto, UpdateProductionDto, CreateProductionDto, TagDto } from "../dto/dto";
-import { NotFoundException, BadRequestException } from "@nestjs/common";
+import { ProductionDto, UpdateProductionDto, CreateProductionDto } from "../dto/dto";
+import { NotFoundException } from "@nestjs/common";
 
 describe("ProductionController", () => {
   let controller: ProductionController;
@@ -20,17 +20,6 @@ describe("ProductionController", () => {
 
   const mockProductions: ProductionDto[] = [mockProduction];
 
-  const mockTags: TagDto[] = [
-    {
-      id: 1,
-      tag: "Drama",
-    },
-    {
-      id: 2,
-      tag: "Classical",
-    },
-  ];
-
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [ProductionController],
@@ -40,16 +29,10 @@ describe("ProductionController", () => {
           useValue: {
             getAllProductions: jest.fn().mockResolvedValue(mockProductions),
             getProductionById: jest.fn().mockResolvedValue(mockProduction),
-            getTagsById: jest.fn().mockResolvedValue(mockTags),
             replaceProduction: jest.fn().mockResolvedValue(mockProduction),
             modifyProduction: jest.fn().mockResolvedValue(mockProduction),
             deleteProduction: jest.fn().mockResolvedValue(undefined),
             createProduction: jest.fn(),
-            getProductionBlogs: jest.fn(),
-            linkBlogToProduction: jest.fn(),
-            unlinkBlogFromProduction: jest.fn(),
-            addTagToProduction: jest.fn(),
-            removeTagFromProduction: jest.fn(),
           },
         },
       ],
@@ -105,10 +88,7 @@ describe("ProductionController", () => {
     });
 
     it("should throw a NotFoundException if the production does not exist", async () => {
-      // Simulate the service not finding the ID in the database
       jest.spyOn(service, "getProductionById").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-    
-      // Verify the controller passes the exact same exception up
       await expect(controller.getProductionById(999)).rejects.toThrow(NotFoundException);
     });
   });
@@ -122,7 +102,6 @@ describe("ProductionController", () => {
 
     it("should throw a NotFoundException if trying to replace a non-existent production", async () => {
       jest.spyOn(service, "replaceProduction").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-    
       await expect(controller.replaceProduction(999, mockProduction)).rejects.toThrow(NotFoundException);
     });
   });
@@ -141,9 +120,7 @@ describe("ProductionController", () => {
 
     it("should throw a NotFoundException if trying to modify a non-existent production", async () => {
       const patchData: UpdateProductionDto = { titel: "A New titel" };
-      
       jest.spyOn(service, "modifyProduction").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-    
       await expect(controller.modifyProduction(999, patchData)).rejects.toThrow(NotFoundException);
     });
   });
@@ -157,141 +134,7 @@ describe("ProductionController", () => {
 
     it("should throw a NotFoundException if trying to delete a non-existent production", async () => {
       jest.spyOn(service, "deleteProduction").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-    
       await expect(controller.deleteProduction(999)).rejects.toThrow(NotFoundException);
-    });
-  });
-
-  describe("-- Blogs --", () => {
-    const mockBlog: BlogDto = {
-      id: 1,
-      titel: "Behind the Scenes",
-      description: "Looking at the set of The Great Show.",
-    };
-
-    describe("getProductionBlogs", () => {
-      it("should return an array of blogs linked to the production", async () => {
-        const expectedBlogs = [mockBlog];
-        jest.spyOn(service, "getProductionBlogs").mockResolvedValueOnce(expectedBlogs);
-
-        const result = await controller.getProductionBlogs(1);
-
-        expect(service.getProductionBlogs).toHaveBeenCalledWith(1);
-        expect(result).toEqual(expectedBlogs);
-      });
-
-      it("should throw a NotFoundException if the production is not found", async () => {
-        jest.spyOn(service, "getProductionBlogs").mockRejectedValueOnce(new NotFoundException());
-
-        await expect(controller.getProductionBlogs(999)).rejects.toThrow(NotFoundException);
-      });
-    });
-
-    describe("linkBlogToProduction", () => {
-      it("should link a blog to a production and return the blog", async () => {
-        jest.spyOn(service, "linkBlogToProduction").mockResolvedValueOnce(mockBlog);
-
-        const result = await controller.linkBlogToProduction(1, 2);
-
-        expect(service.linkBlogToProduction).toHaveBeenCalledWith(1, 2);
-        expect(result).toEqual(mockBlog);
-      });
-
-      it("should pass through errors when linking fails", async () => {
-        jest.spyOn(service, "linkBlogToProduction").mockRejectedValueOnce(new BadRequestException("Already linked"));
-
-        await expect(controller.linkBlogToProduction(1, 2)).rejects.toThrow(BadRequestException);
-      });
-    });
-
-    describe("unlinkBlogFromEvent", () => {
-      it("should unlink a blog from a production and return the production", async () => {
-        jest.spyOn(service, "unlinkBlogFromProduction").mockResolvedValueOnce(mockProduction);
-
-        const result = await controller.unlinkBlogFromEvent(1, 2);
-
-        expect(service.unlinkBlogFromProduction).toHaveBeenCalledWith(1, 2);
-        expect(result).toEqual(mockProduction);
-      });
-
-      it("should pass through errors when unlinking fails", async () => {
-        jest.spyOn(service, "unlinkBlogFromProduction").mockRejectedValueOnce(new NotFoundException("Link not found"));
-
-        await expect(controller.unlinkBlogFromEvent(1, 2)).rejects.toThrow(NotFoundException);
-      });
-    });
-  });
-
-  describe("-- Tags --", () => {
-    describe("getTagsOfProductionByID", () => {
-      it("should return tags for a production", async () => {
-        const result = await controller.getTagsOfProductionByID(1);
-        expect(result).toEqual(mockTags);
-        expect(service.getTagsById).toHaveBeenCalledWith(1);
-      });
-
-      it("should call service.getTagsById with the correct production id", async () => {
-        await controller.getTagsOfProductionByID(1);
-        expect(service.getTagsById).toHaveBeenCalledWith(1);
-        expect(service.getTagsById).toHaveBeenCalledTimes(1);
-      });
-
-      it("should return empty array when production has no tags", async () => {
-        jest.spyOn(service, "getTagsById").mockResolvedValueOnce([]);
-        const result = await controller.getTagsOfProductionByID(1);
-        expect(result).toEqual([]);
-      });
-
-      it("should handle different production ids", async () => {
-        await controller.getTagsOfProductionByID(2);
-        expect(service.getTagsById).toHaveBeenCalledWith(2);
-      });
-
-      it("should throw a NotFoundException if the production does not exist", async () => {
-        jest.spyOn(service, "getTagsById").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-
-        await expect(controller.getTagsOfProductionByID(999)).rejects.toThrow(NotFoundException);
-      });
-
-      it("should handle database errors when fetching tags fails", async () => {
-        jest.spyOn(service, "getTagsById").mockRejectedValueOnce(new Error("Failed to fetch tags"));
-
-        await expect(controller.getTagsOfProductionByID(1)).rejects.toThrow("Failed to fetch tags");
-      });
-    });
-
-    describe("addTagToProduction", () => {
-      it("should add a tag to a production and return the production", async () => {
-        jest.spyOn(service, "addTagToProduction").mockResolvedValueOnce(mockProduction);
-
-        const result = await controller.addTagToProduction(1, 2);
-
-        expect(service.addTagToProduction).toHaveBeenCalledWith(1, 2);
-        expect(result).toEqual(mockProduction);
-      });
-
-      it("should pass through errors if adding a tag fails (e.g., production not found)", async () => {
-        jest.spyOn(service, "addTagToProduction").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-
-        await expect(controller.addTagToProduction(999, 2)).rejects.toThrow(NotFoundException);
-      });
-    });
-
-    describe("removeTagFromProduction", () => {
-      it("should remove a tag from a production and return the production", async () => {
-        jest.spyOn(service, "removeTagFromProduction").mockResolvedValueOnce(mockProduction);
-
-        const result = await controller.removeTagFromProduction(1, 2);
-
-        expect(service.removeTagFromProduction).toHaveBeenCalledWith(1, 2);
-        expect(result).toEqual(mockProduction);
-      });
-
-      it("should pass through errors if removing a tag fails (e.g., production not found)", async () => {
-        jest.spyOn(service, "removeTagFromProduction").mockRejectedValueOnce(new NotFoundException("ProductionDto not found"));
-
-        await expect(controller.removeTagFromProduction(999, 2)).rejects.toThrow(NotFoundException);
-      });
     });
   });
